requirements.txt
Django==1.4.3
coverage==3.6
django-appconf==0.5
django-autoslug==1.6.1
django-compressor==1.2
django-debug-toolbar==0.9.4
django-nose==1.1
gunicorn==0.17.2
ipdb==0.7
ipython==0.13.1
nose==1.2.1
psycopg2==2.4.6
wsgiref==0.1.2
